feat: Save Images as `.webp` (#14)
# Why

With the previous commits reducing the overall footprint of the app, I
aimed to see what else we can do to further this change. One such area
are the images we save to the device for Album, Playlist, and Track
artwork. Images vary in size and adds up fast.


# How

One of the most efficient file formats is `.webp`. Thankfully, Google
has added support for `.webp` images in Android since Android 4 back in
2011 (we support Android versions since Android 6).

To convert images in the initial form as a `base64` image, we utilize
the `expo-image-manipulator` package to convert these `base64` images to
`.webp`.


# Test Plan

We tested the change manually. We observed that:
- We saw a `~30%` decrease in space taken up by images in our app.
- We saw a `~3-4s` increase in the time to save all images in the
initial setup.


# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] Documentation is up to date to reflect these changes (ie:
`CHANGELOG.md` & `README.md`).
- [x] Add new dependencies into `THIRD_PARTY.md`.
- [x] This diff will work correctly for `pnpm android:prod`.